+++ kafka-rado/src/main/java/cz/muni/fi/kafka/rado/FlowProducer.java
@@ -1,4 +1,4 @@
-package cz.muni.fi.kafka;
+package cz.muni.fi.kafka.rado;
 
 import java.io.File;
 import java.util.Properties;
